# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Monitor de Alertas UPP",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/timeline": {
      "get": {
        "description": "Retorna eventos da timeline com filtros opcionais.\nEnriquece os eventos com o nome do paciente.",
        "operationId": "get_timeline_api_timeline_get",
        "parameters": [
          {
            "in": "query",
            "name": "paciente_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Paciente Id"
            }
          },
          {
            "in": "query",
            "name": "tipo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo"
            }
          },
          {
            "in": "query",
            "name": "start_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Ms"
            }
          },
          {
            "in": "query",
            "name": "end_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Ms"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Timeline Api Timeline Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Timeline",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/validate-repositioning/{paciente_id}": {
      "get": {
        "description": "Valida o contrato Backend/Frontend para repouso.\n\nValida:\n1. Último repouso (ultimo_repouso) < Próximo repouso (proximo_repouso)\n2. Próximo repouso (proximo_repouso) > Agora (deve estar no futuro)\n3. Intervalo entre repouso é consistente com perfil\n\nRetorna: {\n    \"valid\": bool,\n    \"errors\": [str],\n    \"ultimo_repouso\": ISO string,\n    \"proximo_repouso\": ISO string,\n    \"intervalo_horas\": float,\n    \"perfil\": str,\n    \"agora\": ISO string\n}",
        "operationId": "validate_repositioning_contract_api_validate_repositioning__paciente_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "paciente_id",
            "required": true,
            "schema": {
              "title": "Paciente Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Validate Repositioning Contract Api Validate Repositioning  Paciente Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Repositioning Contract",
        "tags": [
          "dashboard"
        ]
      }
    },
...
  }
}